move convert_poly_coords to utils.py, change gpkg_layer to layer
